+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/index.js":"1","/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/App.js":"2","/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/ExpenseItem.js":"3","/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/ExpenseDate.js":"4","/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/ExpenseItemsContainer.js":"5","/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/Card.js":"6"},{"size":142,"mtime":1642108551565,"results":"7","hashOfConfig":"8"},{"size":2046,"mtime":1642182147942,"results":"9","hashOfConfig":"8"},{"size":478,"mtime":1642132655928,"results":"10","hashOfConfig":"8"},{"size":493,"mtime":1642131100136,"results":"11","hashOfConfig":"8"},{"size":482,"mtime":1642132704637,"results":"12","hashOfConfig":"8"},{"size":381,"mtime":1642132661209,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"1p0sfjj",{"filePath":"17","messages":"18","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"23","messages":"24","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"25","messages":"26","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/index.js",[],["27","28"],"/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/App.js",[],"/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/ExpenseItem.js",[],"/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/ExpenseDate.js",[],"/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/ExpenseItemsContainer.js",[],"/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/Card.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
+[{"/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/index.js":"1","/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/App.js":"2","/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/Expenses/ExpenseItemsContainer.js":"3","/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/Expenses/ExpenseItem.js":"4","/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/UI/Card.js":"5","/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/Expenses/ExpenseDate.js":"6"},{"size":142,"mtime":1642182511014,"results":"7","hashOfConfig":"8"},{"size":2055,"mtime":1642183184518,"results":"9","hashOfConfig":"8"},{"size":494,"mtime":1642183184540,"results":"10","hashOfConfig":"8"},{"size":490,"mtime":1642183184556,"results":"11","hashOfConfig":"8"},{"size":381,"mtime":1642183133164,"results":"12","hashOfConfig":"8"},{"size":493,"mtime":1642183138439,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1p0sfjj",{"filePath":"16","messages":"17","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/index.js",[],"/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/App.js",[],"/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/Expenses/ExpenseItemsContainer.js",[],"/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/Expenses/ExpenseItem.js",[],"/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/UI/Card.js",[],"/Users/tomtom/Documents/Coding/Udemy/React_Max/Section_3 React Bascs/src/components/Expenses/ExpenseDate.js",[]]
