+++ src/components/Expenses/ExpenseItem.js
@@ -1,4 +1,4 @@
-import Card from "./Card";
+import Card from "components/UI/Card";
 import ExpenseDate from "./ExpenseDate";
 import "./ExpenseItem.css";
